+++ exo-go/src/applicationrunner/service_restarter.go
@@ -1,4 +1,4 @@
-package runner
+package applicationrunner
 
 import (
 	"fmt"
